Added files via download
I really enjoyed this simple yet quite challenging project as it really caused me to think and really be a problem solver. For example, when I was figuring out how to try and put the cookie image over the ball so you could only click the cookie to improve your click score that was quite challenging and I had to try many different things until I eventually settled on image sprite!