Update README.md
Added my 2 sentences swell as my two screenshots to the read me and just incase adding the 2 sentences here! 

What was quite difficult about this project was the problem solving it required as it really forced you to try different things and try new methods to create the design you wanted. For example, this happened to me especially not on the code as that was given to us but the design view as when I was trying to echance the image of the ball and make it a cookie I had to try many different image processes until I finally settled on my desired method!